Coding Standards: Cast `gmdate( 'Z' )` to an integer before addition.
This addresses two instances of the (numeric string) `gmdate( 'Z' )` being added to an `int` value.

Affected functions:
* `upgrade_110()`
* `WP_Date_Query::validate_date_values()`

Follow-up to [942], [29925], [45424].

Props justlevine.
See #52217.
Built from https://develop.svn.wordpress.org/trunk@59465


git-svn-id: http://core.svn.wordpress.org/trunk@58851 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/includes/upgrade.php
@@ -992,7 +992,7 @@ function upgrade_110() {
 
 	$time_difference = $all_options->time_difference;
 
-	$server_time    = time() + gmdate( 'Z' );
+	$server_time    = time() + (int) gmdate( 'Z' );
 	$weblogger_time = $server_time + $time_difference * HOUR_IN_SECONDS;
 	$gmt_time       = time();
